Rethought interface state.
A component with interface state (which should be in redux, everyone
else is wrong), should manage defaults for that state itself, either
with props, defaultProps, or both, and register itself with the store.

My previous approach of defining the initial application state in
reducers was reasonably inextensible, requiring a change to include
injecting initial state into the reducer when creating the store for
every component's interface state.